requirements.txt
asn1crypto==0.24.0
attrdict==2.0.1
attrs==17.4.0
Automat==0.6.0
base58==1.0.3
blinker==1.4
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
cloud-init==19.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cytoolz==0.10.0
distro-info===0.18ubuntu0.18.04.1
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.0
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.1
grpcio==1.22.0
grpcio-tools==1.22.0
hexbytes==0.2.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
parsimonious==0.8.1
protobuf==3.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
rlp==1.1.0
secp256k1==0.13.2
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
ssh-import-id==5.7
systemd-python==234
toolz==0.10.0
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
zope.interface==4.3.2
